`then` only can return `Thenable`
It is with heavy heart I get here. But I’ve seen Swift do things like infer:

	foo().then {
		Promise(something)
	}.then {
		XCTAssert($0 is Promise<Promise<something>>)
    }

This is usually caught be the developer. Because they check the type of the subsequent closure somehow.

But I found a case where I didn’t and it truly fucked up my app. I was expecting some extremely important promise to run, and obviously, it did not, because it was wrapped in another promise.

So we no longer support returning plain T from `then`. I see no way around it. In many ways this is more fitting to Swift, so there we go.