`Promise { seal` -> `Promise(.pending) { seal`
Because otherwise in some cases Swift infers the closure to`Promise(value:)` resulting in `Promise<(Resolver) -> Void>` which we obviously never want.

In general we would hope that this would be detected quickly by the user, but I have found for short chains you can often not realize, which can lead to severe bugs.

So, we must be explicit.